+++ doc/release-notes.md
@@ -4,4 +4,5 @@ Bitcoin ABC version 0.26.10 is now available from:
 
   <https://download.bitcoinabc.org/0.26.10/>
 
-This is a maintenance release with no user-visible change.
+ - The `getavalancheinfo` RPC returns a new `verification_status` field
+   with a status description string to indicate why local proof is not verified.